Google Science Fair 2018/requirements.txt
imutils==0.5.2
numpy==1.15.4
opencv-python==3.4.4.19
Pillow==5.3.0